Major Changes
Due to Issues with httprouter, I decided to switch back to mux.
I have also switched from glide to Godep for managing
dependencies. I have also updated the badge so that it looks
better on High DPI displays (the Hackaday logo is now an SVG, not
just an image embedded in an SVG). I have also replaced the classic
"Fork me on GitHub" ribbon with a GitHub corner. I also updated
jQuery, Bootstrap, and Prism.